Develop

+++ index.js
@@ -36,6 +36,10 @@ window.onload = () => {
       this.list.push(todo);
     }
 
+    remove({ todoId }) {
+      this.list = this.list.filter(({ id }) => id !== todoId);
+    }
+
     updateTodo({ todoId, values }) {
       const targetTodo = this.list.find(({ id }) => id === todoId);
       targetTodo.set(values);
@@ -58,18 +62,23 @@ window.onload = () => {
     completeAt;
     isDone;
     priority;
+    parentId;
     
-    constructor({ id, title, content = '', createdAt = new Date(), isDone = false, completeAt, priority = PRIORITY.MEDIUM }) {
+    constructor({ id, title, content = '', createdAt = new Date(), isDone = false, completeAt, order = 1, priority = PRIORITY.MEDIUM, parentId }) {
       if (typeof title === 'undefined') {
         throw new Error('Title is required');
       }
-      this.id = id || Date.now();
+      this.id = isNumberString(id) ? id : Date.now().toString();
       this.title = title;
       this.content = content;
       this.createdAt = createdAt;
       this.completeAt = completeAt;
       this.isDone = isDone;
       this.priority = priority;
+      this.order = isNumberString(order) ? parseInt(order) : 1;
+      if (isNumberString(parentId)) {
+        this.parentId = parentId;
+      }
     }
 
     set(values) {
@@ -85,6 +94,7 @@ window.onload = () => {
 
 
   // functions
+  const isNumberString = (id) => !isNaN(parseInt(id));
   const checkFormValidation = ({ data: { title = '' }}) => {
     if (title.length === 0) {
       return {key: 'title', message: 'title is required'};
@@ -99,16 +109,15 @@ window.onload = () => {
   const onEditTodo = ({ target }) => {
     const { dataset: { id }} = target.closest('.todo');
     const { value, dataset: {key} } = target;
-    todoList.updateTodo({ todoId: parseInt(id), values: {[key]: value }});
+    todoList.updateTodo({ todoId: id, values: {[key]: value }});
     updateStorage({ list: todoList.list });
   }
 
   const getRandomColor = () => Object.keys(COLOR)[Math.floor(Math.random(0, 12) * 10)];
-  const formatDate = (date) => `${date.getFullYear()}-${date.getMonth()}-${date.getDate()}`;
  
   const renderTodo = ({ id, title, content, completeAt, isDone, priority }) => {
     return `
-      <li class="card list__item todo" data-id=${id}>
+      <li class="card list__item todo" data-id=${id} draggable=true>
         <span class="card__label ${getRandomColor()}"></span>
         <div class="todo__body">
           <h5 class="todo__title">
@@ -116,15 +125,11 @@ window.onload = () => {
           </h5>
           <div class="row">
             <span class="todo__date">
-              ${
-                completeAt
-                ? `<input 
-                    class="content__complete-at"
-                    type="date"
-                    data-key="completeAt"
-                    value="${new Date(completeAt).toISOString().slice(0, 10)}">`
-                : ''
-              }
+            <input 
+              class="content__complete-at"
+              type="date"
+              data-key="completeAt"
+              value="${completeAt && new Date(completeAt).toISOString().slice(0, 10)}">
             </span>
             <span class=""todo__priority>
               <select
@@ -137,22 +142,22 @@ window.onload = () => {
                 </option>
                 <option
                   value ="${PRIORITY.HIGH}"
-                  ${priority === PRIORITY.HIGH ? 'selected' : 'none'}>
+                  ${priority === PRIORITY.HIGH ? 'selected' : ''}>
                   High
                 </option>
                 <option
                   value ="${PRIORITY.MEDIUM}"
-                  ${priority === PRIORITY.MEDIUM ? 'selected' : 'none'}>
+                  ${priority === PRIORITY.MEDIUM ? 'selected' : ''}>
                   Medium
                 </option>
                 <option
                   value ="${PRIORITY.LOW}"
-                  ${priority === PRIORITY.LOW ? 'selected' : 'none'}>
+                  ${priority === PRIORITY.LOW ? 'selected' : ''}>
                   Low
                 </option>
                 <option
                   value ="${PRIORITY.LOWEST}"
-                  ${priority === PRIORITY.LOWEST ? 'selected' : 'none'}>
+                  ${priority === PRIORITY.LOWEST ? 'selected' : ''}>
                   Lowest
                 </option>
               </select>
@@ -161,40 +166,84 @@ window.onload = () => {
           <p class="todo__content">
             <input class="content__input" data-key="content" value="${content}">
           </p>
-          ${
-            isDone
-              ? `<button class="todo__done todo__done--clear" data-key="isDone">Incomplete</button>`
-              : `<button class="todo__done" data-key="isDone">Complete</button>`
-          }
+          <div class="todo__buttons row">
+            <button class="todo__delete">Delete</button>
+            ${
+              isDone
+                ? `<button class="todo__done todo__done--clear">Incomplete</button>`
+                : `<button class="todo__done">Complete</button>`
+            }
+          </div>
         </div>
       </li>
     `
   }
 
+  const renderTodos = ({ todoList, parentId }) => {
+    if (!todoList?.list) {
+      return;
+    }
+    const { list } = todoList;
+    const todoListDomString = list
+      .filter((todo) => todo?.id && parentId === todo.parentId)
+      .sort((prev, next) => prev.order - next.order)
+      .map(renderTodo)
+      .join('');
+    $todoList.innerHTML = todoListDomString;
+
+    $todoList.dataset.id = parentId;
+  
+    $moveParent.classList[parentId ? 'add' : 'remove']('move-parent__visible');
+
+    const parentTodo = list.find(({ id }) => id === parentId);
+    $parentName.innerText = parentTodo ? `${parentTodo.title}'s ` : '';
+  }
+
+  const resetFields = () => {
+    const formInputs = $form.querySelectorAll('.form__input');
+    for (const $input of formInputs) {
+      if ($input.id === 'priority') {
+        $input.value = PRIORITY.MEDIUM;
+        return;
+      }
+      $input.value = '';
+    }
+  }
+
+
+
   // Dom elements
   const $app = document.querySelector('.todo-app');
   const $form = $app.querySelector('.todo-app__form');
+  const $moveParent = $app.querySelector('.move-parent');
+  const $parentName = $app.querySelector('.parent-name');
   const $todoList = $app.querySelector('.todo-app__list');
 
 
 
   // Initialize
+  // loading
+  setTimeout(() => {
+    $app.classList.remove('todo-app__loading');
+  }, 3000);
   const storageList = JSON.parse(localStorage.getItem(LOCAL_STORAGE_KEY)) || [];
   const todoList = new TodoList({ list: storageList });
-  const todoListDomString = todoList.list
-    .filter((todo) => todo && todo.id)
-    .map(renderTodo)
-    .join('');
-  $todoList.insertAdjacentHTML('beforeend', todoListDomString);
+  renderTodos({ todoList });
+
 
 
   // Event
   $form.addEventListener('submit', (event) => {
     event.preventDefault();
+    for (const $error of $form.querySelectorAll('.form__error')) {
+      $error.remove();
+    }
+  
     const { currentTarget } = event;
     const $items = currentTarget.querySelectorAll('.form__item');
     const formEntries = Array.prototype.map.call($items, ($item) => {
-      const {value = '', dataset: { key }} = $item.querySelector('input, select, textarea');
+      const {dataset: { key }} = $item;
+      const {value = ''} = $item.querySelector('input, select, textarea');
       return [key, value];
     });
     
@@ -209,7 +258,12 @@ window.onload = () => {
       `);
       return;
     }
-    const newTodo = new Todo(formData);
+  
+    if(!confirm('Want to add?')) {
+      return;
+    }
+    resetFields();
+    const newTodo = new Todo({ ...formData, parentId: $todoList.dataset.id });
     todoList.add({ todo: newTodo });
     localStorage.setItem(
       LOCAL_STORAGE_KEY,
@@ -221,6 +275,7 @@ window.onload = () => {
     );
   });
 
+  // edit
   $todoList.addEventListener('focusout', (event) => {
     const { target } = event;
     if (target.dataset.key && target.tagName !== 'BUTTON') {
@@ -228,6 +283,7 @@ window.onload = () => {
     }
   });
 
+  // edit
   $todoList.addEventListener('keydown', (event) => {
     const { target, target: { dataset: { key } }, code } = event;
     if (key && code === 'Enter') {
@@ -237,17 +293,66 @@ window.onload = () => {
   });
 
   $todoList.addEventListener('click', ({ target }) => {
-    const { dataset: { key }} = target;
-    const { dataset: { id }} = target.closest('.todo');
-    if (key === 'isDone') {
+    const $todo = target.closest('.todo');
+    if (!$todo) {
+      return;
+    }
+    const { dataset: { id: todoId }} = $todo;
+    // toggle done
+    if (target.classList.contains('todo__done')) {
       const isDone = target.classList.contains('todo__done--clear');
       todoList.updateTodo({
-        todoId: parseInt(id),
+        todoId,
         values: { isDone: !isDone }
       });
       updateStorage({ list: todoList.list });
       target.innerText = isDone ? 'Complete' : 'Incomplete';
       (target.classList[isDone ? 'remove' : 'add'])('todo__done--clear');
+      return;
+    }
+    // delete todo
+    if (target.classList.contains('todo__delete') && confirm('Want to delete?')) {
+      todoList.remove({ todoId });
+      updateStorage({ list: todoList.list });
+      $todo.remove();
+      return;
+    }
+    // move
+    if (target.classList.contains('todo')) {
+      renderTodos({ todoList: todoList, parentId: todoId });
+      return;
     }
   });
+
+  // drag & sort
+  $todoList.addEventListener('drag', (event) => {
+    const {target, clientX, clientY} = event;
+    let swapItem = document.elementFromPoint(clientX, clientY) || target;
+    if (swapItem.parentNode === $todoList) {
+      swapItem = swapItem !== target.nextSibling
+        ? swapItem
+        : swapItem.nextSibling;
+      $todoList.insertBefore(target, swapItem);
+    }
+  });
+
+  $todoList.addEventListener('dragover', (event) => {
+    event.preventDefault();
+  }, false);
+
+  $todoList.addEventListener('drop', () => {
+    const todos = $todoList.querySelectorAll('li');
+    for (let index = 0; index < todos.length; index++) {
+      const {dataset: {id}} = todos[index];
+      todoList.updateTodo({ todoId: id, values: {order: index + 1}});
+    }
+    updateStorage({ list: todoList.list });
+  });
+
+  // move to parent
+  $moveParent.addEventListener('click', () => {
+    const {parentId} = todoList.list.find(({ id }) => id === $todoList.dataset.id);
+    renderTodos({ todoList: todoList, parentId });
+  });
 }
+
